Fix list item clickability.
Made it so that pointer is clickable when hovering over unexpanded list items, and only up chevron can close list items. Also, a elements that may be visible on unexpanded and faded list item aren't clickable until the list item is expanded. Other small changes: updated the expand and collapse icons to be larger chevrons; added a bit of padding between list items for better spacing on smaller mobile devices; updated a bit of the wording in index.html