new front page & footer content.
this is a big(ger) change, but was initially prompted by desire to add a PGP link to the footer, & realizing that there was no good "icon" that would fit it. this led me down rabbit hole of redesigning my entire footer to be more practical, & I took opportunity to also update `index.html`.

summary of changes:
  - removed comments in `index.html` & `_includes/footer.html`, & cleaned up these files
  - created new `_includes/external.html` snippet to use for external links
  - added profile picture
  - update CSS to center everything more nicely on page, & reduce width to 450px. was inspired by https://kennandavison.com/ for setting width of main content container & for profile pic dimensions. stuff is left-aligned now.
  - delete TODOs in `README.md`