cleanup CSS file TODOs
mostly getting rid of no longer needed sections that used to control UX of list items that I had on front page, but I got rid of those a while ago. also, realized that in my main page template was using both a "container" & a "section" class as wrappers for content, so consolidated everything into just "container". Because this is using `display: flex`, it required a few small changes to get the exact same UI as before, but this should be simpler to think through moving forward.

also removed all corresponding mentions of deleted stuff from JS file

functionally, the only difference now is that `text-align: justify` is not set, which means that <p> elements won't add spacing just to fit width of container, which I think looks better.